Replace ClassJarProvider#findModuleClassFile with project system API
Currently, when we search for class files to load during layout
rendering we search through each module in the transitive dependencies
of the module containing the layout. For each module, we check the JPS
output directory for the class file and if that doesn't work we use
a build system-specific strategy defined by
ClassJarProvider#findModuleClassFile.

Instead, we'd like for the entire process of finding a class file to be
customizable based on the build system. This would allow us, for
example, to consider the entire module transitive closure for Blaze
projects rather than searching the modules one at a time.

This CL enables this kind of customization by replacing the
findModuleClassFile API with AndroidModuleSystem#findClassFile,
which searches through a module and its dependencies according to
some build system-specific strategy.

This CL simply reorganizes the existing logic for each build system,
so class loading should behave exactly as it did before.

Bug: 71713776
Test: ModuleClassLoaderTest
      + manually tested rendering layouts with custom views
        in a Gradle project

Change-Id: Iec5e4dd01222f9cfc3180f50c201e914c3c14057