Add disk space analysis to SDK manager operations dialog.
This CL improves SDK operations confirmation dialog by
adding disk space analysis in relation to the expected
download size and the space available on the SDK root
partition and the temp partition.

Except for giving the user a better idea of disk utilisation,
it also issues a warning if there is a risk of operation being
not completed successfully due to disk space issues.

Test: SdkUpdaterConfigurable doesn't have a test class at present,
producing one is TBD.
Bug: b/72581483, b/79409651

Change-Id: I9deea3dc57ebadd3ff32a589ed1484f91b5957aa