Set valid initial spinner index in RecyclerViewAssistant
Currently the constructor in RecyclerViewAssistant checks the list of
all templates to see if one matches the sampledata of the RecyclerView.
If one does, then it sets the initial index of its spinner to the
corresponding index of the template in the list. The spinner's model
is actually a filtered version of that list, so this can lead to
IndexOutOfBoundsExceptions.

This CL searches through the templates in the spinner's model instead
of the list of all possible templates, ensuring that the index is always
valid.

Fixes: 111407750
Test: Manually confirmed that you can start on the last template
      and this doesn't cause an IndexOutOfBoundsException.

Change-Id: I40a8def9ef23a9a684b6db81d42b54bfb1714be9