Filter sampledata templates by available dependencies
Currently, RecyclerViewAssistant only filters out which sampledata
templates it shows to the user by whether or not the module depends
on androidx. This means that the Email template, which uses constraint
layout, is always available to the user, even if the module doesn't
depend on constraint layout and the template won't render correctly.

This CL updates RecyclerViewAssistant to filter templates for which
the module doesn't have the necessary dependencies.

Fixes: 111433513
Test: confirmed Email template only available when module depends
      on constraint layout

Change-Id: Icc0a4339f69d4b41ab31bcdcd010e88bc2c50229