Track user sentiment
Users who have opted in to helping improve the product by sending
Google metrics, will now get two icons in the status bar
to indicate positive or negative sentiment towards the product.

Once they click these icons, feedback is sent to Google and
a notification is given thanking them for their feedback.

In the case of negative sentiment, the user is given the option
to report a bug.

Icons are material design icons from https://material.io/tools/icons/
Sized to 16x16 (32x32 for high-dpi) and to use the IDE's color scheme
(#808080 on transparent) to work on light & dark themes.

Bug: 110560319
Test: manual
Change-Id: I46f788d096a2360c32915e0d3a8469cadc59a9e2