Add 'Hide Obsolete Packages' functionality to SDK manager.
This CL introduces a new checkbox for hiding obsolete packages,
both on Platforms and Tools tabs of SDK manager. It is selected
by default and leads to a less cluttered packages view, both
in summarized and detailed modes.

Apart from that this CL polishes the detailed platform view by
ensuring no empty platform nodes are shown in the UI.

Test: These panels do not have dedicated test classes, producing
them is TBD. In the mean time, verified visually that UI WAI.
Bug: b/111504937

Change-Id: I1fa6abee67baa945e725cda51c39fb005f0411b1