Notify sync finished when error in post sync setup
Exceptions must be caught and sync finished should be reported to avoid
sync task not finishing correctly (e.g. indexing suspender running
forever as it is never notified that sync finished).

Test: PostSyncProjectSetupTest.testWithExceptionDuringProjectSetup and
manually by throwing a runtime exception in any part of setup and make
sure sync is not stuck forever in this case.
Fixes: 111432735

Change-Id: I465c43bf8110e3da788f7074bdab1202b399c2cb