Updating AtraceFrame code to align with studio frames feature.
This change removes a bunch of the logic to associate Main thread
events with render thread events. This logic has been removed and
simplified as we are no longer going down the path of automagically
doing things for our users. We want users to be able to define what
threads/events they care about tracking and allow them to track those
events in the frames bar.

While this change has no UI changes it does set us up to allow us to
track the main thread, and render thread events independently to provide
a similar experience to what the previous frame code was doing.

Test: Updated, AtraceFrameTest + FactoryTest, added Filter Test.
Bug: 74404740
Change-Id: I43786bff5ea4147fe4f626e51f2568f1165504f1