Fix capture details panel closing on code navigation
When we jump to source using one of the charts/trees of the capture details
panel, we set the profiler mode to NORMAL to unmaximize the profiler tool
window so the user can see a portion of the code editor and inspect the code
they've just navigated to.

The problem was we were hiding the capture details panel when we did that,
which should not happen. When we get to L3, we shouldn't go back to L2
anymore.

Bug: 111551264
Test: CpuProfilerStageViewTest#dontHideDetailsPanelWhenGoingBackToNormalMode
Change-Id: Ifb00b491f7af4a5d5c013e4e133561de365d5827