Disable the "generated file" notification for APK editors
When opening the APK editor, specifically for Android App Bundle files,
we don't want to show the notification about not editing files in the
generated folder.

To avoid adding a dependency between the notification provider and
the apk file editor modules, we use a new Key defined in the
"android-common" module.

Fixes: 111696015
Test: Included
Change-Id: Iffebfe152c595bfa44d229472c652155c29dc406