Remove MemoryService's dependency on ProfilerTable
This allows us to remove a custom, one-off initialization method that
exists only for this particular case; instead, we create service stubs
directly, which is what other services do.

This change on its own may seem unecessary, but it will help with a
testing CL being worked on separately, which expects all methods in a
table to be relevant to querying / updating the table.

Bug: N/A
Test: Refactoring only, should be covered by existing tests
Change-Id: I4c4fba4532396c93142d234916b8da7f0406427b