Add RenderPropertiesAccessManager extension point
Determining the value of Blaze experiments requires system properties
access, which isn't allowed by RenderSecurityManager. This causes a
render sandboxing error when loading any layout with custom views in
ASWB.

This CL resolves the issue by creating the
RenderSecurityManagerOverrides extension point and having
RenderSecurityManager delegate to the extensions to determine whether
system properties access is allowed. Access is opt-in, and we'll only
provide an implementation for the ASWB plugin, so this should only
effect ASWB.

Bug: 111519191
Test: Verified Gradle layouts render normally
Change-Id: Ic063cd55a98e1cce22cc28096f5c2a929d1ed6d2