Remove PARSING(_FAILURE) from CPU profiler state machine
As part of the CPU profiler state machine refactoring, we're removing
the PARSING and PARSING_FAILURE states from it. After this change, the
states remaining on the state machine are: IDLE, STARTING, CAPTURING, and
STOPPING.

Now, when parsing starts, we fire CpuProfilerAspect.CAPTURE_PARSING, which
is a dedicated aspect to notify listeners about parsing, which should be
orthogonal to capture state. The listeners of the aspect can call
CpuCaptureParser#isParsing to check wether parsing is in progress.

Bug: 111150213
Test: updated affected tests
Change-Id: I95c15508653dbac502186fc09a6cd590d772fb63