Fix the library name mismatch in AarResourceRepositoryCache
- Normalize the IntelliJ library names created by Gradle sync to be
"Gradle: " + library.getArtifactAddress() for both AARs and JARs.
- When we don't have a model to work with, when getting the AAR
repository from the cache, strip this "Gradle: " prefix from the
IntelliJ library name to get the libraryName that the repository should
use. This will now match if the same exploded folder is used with and
without the model by the same IDE process.
- Separately I will look into why we may end up in the model-less path
when working with Studio+Gradle projects. This may be about timing, e.g.
trying to use resources before the sync is done.

Test: Manually so far. One way to repro is to create a new project, edit
build.gradle but don't sync, switch to MainActibity.java, close the
project and open it again from the "recent projects" list. If we agree
on the approach to fix the problem, I will try to automate it.
Bug: 79942260
Change-Id: I86fc95a6bfc1f4e68e7e4ec514494625ee093ca3