Ensure that re-sync prompt is shown when CMake file is changed.
This change is not yet all the way there.  If the user starts on the
CMakeLists.txt file and changes that, the notification doesn't show.
I believe that's because the file is not yet known to be an external
build file and thus the tree change listener is not set for it in line
98.  I'm not currently sure how to fix that.  Maybe we could add a
heuristic to the isExternalBuildFile function that either
CMakeLists.txt or Android.mk should return true if we're not done
parsing the gradle files yet.

Change-Id: Ifb455c5a70702c15c531fd55be5ab976ff20ad6e
Fixes: 70209187
Test: Manual