Updating argument editor behavior
- When we click on an action pointing to a navigation, the arguments
shown in the inspector should be those of the start destination of the
navigation.
- When we click on a navigation, the arguments shown in the inspector
should be those of the start destination, and we should only be allowed
to edit the default values (similar to an action).
- When we change the start destination of a navigation, the argument
default values should be cleared.

Bug: 111757065 (Note that this does not completely fix that bug, as the
case of arguments on includes is not handled).
Test: Added tests in NavArgumentDefaultValuesPropertyTest, updated tests
in NavInspectorPanelTest and SetStartDestinationPropertyTest.

Change-Id: I4d3a37aeb8c483786b1ff7340324e248bee86c4a