[New PSD] Reflect live model collection changes onto UI.
Gradle project models are resolved in the background. When the process
completes the UI models are updated and undeclared
entities like build type "debug" may be added to the collections. The
UI should be updated to avoid users' attempts to re-create them.

This CL makes the UI to be automatically refreshed when its
PsModelCollection changes.

Fixes: 111822303
Test: unit tests updated
Change-Id: I34f2354fdb7c65e4bf1a74fe58c335a02dacd18f