Adjust timeout rules in GuiTestRule
This reduces the (probably excessive) timeout in IdeControl from 2
minutes to 1 minute, and increases the difference between inner and
outer timeouts in GuiTestRule to 2 minutes. This should make it
substantially less likely for the before-test portion of the Rule code
to take long enough that it prevents the after-test Rule code from
getting a chance to run.

Also move LogStartAndStop outside the outer timeout, so that the test
result is logged even if the outer timeout triggers.

Change-Id: Ic74b3acae14872908c850d88b093c455a66874e0
Fixes: 111861465