[New PSD] Avoid expensive UI tree re-initialization
MasterDetailsComponent.updateSelection() is called when the user makes a
selection in the tree and to initialize the selection in response to
addNotify() when is being added to the element tree. It always removes
the configuration panel from the element tree and adds it back. It might
be an expensive operation if EditorTextField is involved. However, if
MasterDetailsComponent are nested the number of addNotify()
increases exponentially.

Bug: n/a
Test: n/a (manually tested, future UI tests)
Change-Id: Idd70bc9a592ff3a5a166fa2a430513852d39ae62