Rewrite NavigationSchema
NavigationSchema is now structured around two mappings:
- Tag name to destination class
- Destination type to destination class

Both are in theory many-to-many mappings, though in practice they may
often be one-to-one (since the type to class map may only contain the
roots of the destination class hierarchies).

The class provides two main areas of functionality:
- mapping between tag name, destination type, and destination class
- providing information on the xml schema of navigation files, including
  valid root tags and valid children for a given tag.

(It also currently provides some UI-related information, but that should
be moved out in a subsequent CL).

Test: updated existing tests. Once the destination class annotation is
available in the library (already checked in for the next release) we'll
be able to add more interesting tests.
Bug: 110763342
Change-Id: I90f87892da9ea26c8a1353a38aa5ef685bbea28f