studio stats: clean up AnalyticsSettings API
AnalyticsSettings API used to be instance based, this was exposing an
implementation detail of the GSON object being serialized.

This caused issues as people kept cached instances around.

This change makes all public API static on AnalyticsSettings
and creats a separate AnalyticsSettingsData object that
is only used for serialization and testing.

Change-Id: I7c928ac72ed710270e4f7f806b2f5016ab53df4e
Test: updated
Bug: 110560339