Adding FakeTimer to StudioProfilers in Test.
Without this change we end up leaking a studio profilers instance
in the test causing chaos when all the test are run at once.
This was detected as a result of the test timeout for
DataSeriesPerformance test.

This change fixes the issue by not using a threaded timer to tick the
studio timer. As such studio profilers is cleaned up after each test run
because the JVM doesn't have an active timer running. This only fixes
the test, we need to address the attached bug to fix the issue in studio.

Test: Updated
Bug: 112055245
Change-Id: I219cd5b9bb3833ecb8bbcaa8eb5a83798e3ceea6