Reimplements deploy metrics to be one object
Instead of logging all events separately, this change tracks
the run cycle with one object that is sent at the end.

In order to make sure the event is aborted or sent, the service
still remains and tracks the last object.

Also, if the stats object is for some reason not present in the
current execution environment, a "partial" deploy is started from
that point.

Test: to follow up with deploy tests.
Bug: 111266263
Change-Id: I4aad6f66349cf4efa21b1d969d628bc0aa38de67