Add Swing tree serialization.
This CL adds path-to-root serialization with respect to the Swing component
hierarchy.

Note that each node only serializes a path fragment from itself to the last
parent component that was also captured by swingp (there are exceptions to this,
such as JComponents rendered by the InstructionsRenderer). Therefore, the
whole swingp hierarchy is needed to determine the hierarchy of any individual
JComponent that was captured by swingp.

This CL also added example filtering using serialization to visualizer.

Bug: 80499904
Change-Id: Iabf7a21a0e0355e334ef4b61d0aae22c529f43b6
Test: Added tests.