Display variants in "Generate Singed Bundle" window.
In the current window, there are two separate selections for
build type and flavors. Users are allowed to select one build
type and one or more flavors. This CL combines these two fields
and display one full list of variants. Users are allowed to
select one or more variants.

Making the change because user works with variants in the IDE,
also the code of Gradle Sync uses variants as identifiers. This
avoids the work of spliting debug type and flavor from variant name
as well as combining them together to get the variant.

Bug: 111929608
Test: ExportSignedPackageTest

Change-Id: Ifccb93f0a0975b8a56d2289aa623f8b7b1039a5d