Make browseButton an optional property.
This is a refactoring only:
Before this change an implementaion of PropertyItem could also
implement a ActionButtonSupport interface to get a browse button
to the right of the property editor.

After this change the browse button is specified as a property of
PropertyItem.

Do this in preparation for adding another ActionButton for color
controls in an upcoming CL.

Bug: 73778668
Test: Existing tests should pass (this is a refactoring).

Change-Id: I300ebc43568d0ed8cf17f42a6068269e8ce09a11