Run emulator within the Bazel sandbox for UI tests
Some UI tests use the emulator. Providing a fake emulator to mock out
the emulator from these tests would be prohibitively expensive to do, as
it would require the team to write a new lightweight emulator instead.
This change instead introduces a TestRule and a single test -
DebugOnEmulatorTest - to demonstrate how to launch and use the emulator
when running within a Bazel sandbox.

Test: bazel test //tools/adt/idea/android-uitests:DebugOnEmulatorTest
Bug: 112553978
Change-Id: I83ee44b505f5e5807a4d90b9dab2c3008530b35a