Separated checkboxes and texts to align description
Separated checkboxes and texts into different objects and put them
in different columns in the grid to be able to align description
with the checkbox text. The problem is that there is a significant
gap between the checkbox and the text that is about 5 times the normal
gap between text and the checkbox in JCheckbox that appears to be not
controllable from the form file. So when the gap is clicked,
the checkbox is not switched.

Bug: 68986525
Test: From edit run configurations select the profiling tab, text should be
aligned, clicking the text of the checkbox should switch it, however
clicking the gap between the checkbox and the text don't.
Change-Id: Iac9da5c7527acb3d35fdc1b90d191fecf7993616