[SVS] Support of native modules in BuildAction.
In build action, request NativeAndroidProject with parameterized
APIs, then request NativeVariantAbi separately with variant name
and abi name as parameter.

Update SelectedVariant and ModuleDependency to include a nullable
field of abi name.
Update VariantChooser to choose abi based on user selection or
x86 if it's the first sync.

The abi of dependency modules are not sepecified by dependent module,
always try to use the same abi name as the dependent module.
And if it doesn't exist, then pick x86 or the first one from
available abi list.

Bug: 79581802
Test: Integration tests in a separate CL.
Change-Id: I275b762d276840bbb6fce829c20e9cb2f41b6161