[AndroidX] Remove dumb mode in sync after androidx/appcompat refactor
This is causing issues within DumbServiceImpl and ProgressSuspender
where processes are sometimes being suspended and never resumed
leading to permanent indexing (or VCS processes).

I believe this has to do with the interactions between running and dumb
mode task and the IndexingSuspender that is run as a result of sync.

Change-Id: Ib60f28e2743ed0f794c9057c48d5b86ac8d656e6
Fixes: 112536288
Test: N/A