Hide special-case data viewer logic as an implementation detail
A first pass at adding code related to previewing data
as both raw and pretty-printed (e.g. relevant for JSON,
XML, etc.) some special-case code ended up living in the
network profiler, instead of inside the data viewer code
itself.

At the time of the review, it was decided that the logic
shouldn't live there, but because of time constraints,
the cleanup should happen in a followup CL... which is
this one!

The only logic change is we now use a TextPane to
display raw text instead of an IntelliJ editor
in all cases (rather than using special case logic
to only do the replacement for a handful of file
types).

Test: N/A, this CL deletes and shuffles existing code.
his isis only a shuffling of (and deleting of)
existing code.

Fixes: 111835527

Change-Id: I372e251715eb5f6a85314a20e57b7a3952abe110