Fix zoom issue in profiler timeline by capping mouse wheel rotation
On mac, when you flick the mouse wheel you are very likely to encounter a rotation count > 8. We use the count * zoom factor (0.125) to determine how much to zoom in by, and as you can see when count > 8 you are zooming in more than 100% which doesn't make sense.

We fix it by limiting the delta of the zoom range to 90% of the current view range in zoom().

Change-Id: I8a842a93c9a0c9be94a4e29db96ae214679c4199
Fixes: b/109946020
Test: Added new test to check for the protection of zoom() against mouse wheel generated delta ranges that exceed current view range.