Make a SceneComponent for Actions
Previously there was no SceneComponent created for normal (or
dotted-line) actions: they were represented as targets on the source
destination. Now they are SceneComponents. It doesn't make too much
sense for them to be children of a specific destination, so instead
they're children of the root. Selection is still handled by
ActionTarget (ideally all selection would be handled only by targets).
Notably, this means that there are currently no ill effects of the fact
that we don't set position or bounds for the component itself (since
drawing is based entirely on the position of other components).

Test: updated existing tests
Bug: related to 80325197
Change-Id: I9bb06e0ad802013b9efc4c7729507486152b6d95