Moved the total field in Memory tooltip to match the specs
Moved the total memory field in the memory tooltip to the bottom to match
the specs. The separator between the total memory field and the other fields
in the specs is not possible to be added with the current design of the
legends, so this is not a complete fix to the bug.

Bug: 79220099
Test: Run MemoryProfilerStageTest. For manual testing, In the memory
profiler hover over the memory line charts at any specific time, a tooltip
should show up with the total memory field in the bottom.

Change-Id: Ief3c5e4eb595eb2eed8cd50c185b04ca1d0b3207