Delegate to module system to choose sample data location
Currently, SampleDataResourceRepository#getSampleDataDir creates
a module's sample data directory in the main content root of the module.
In a Gradle project, this means that the directory will sit parallel
to the module's resources. For Blaze projects, however, the main content
root of a resource module is the resource folder iteself. This causes
Blaze sync issues when the user creates a sample data directory.

This CL helps resolve the issue by migrating the API to
AndroidModuleSystem#getSampleDataDirectory, which allows us to do
something different when creating the sample data directory for a
resource module in a Blaze project. This is accomplished by introducing
the SampleDataDirectoryProviderInterface. GradleModuleSystem and
DefaultModuleSystem delegate to the default implementation in
MainContentRootSampleDataDirectoryProvider, which preserves the existing
behavior of putting the sample data directory in the main content root.

Bug: 112353103
Test: MainContentRootSampleDataDirectoryProviderTest

Change-Id: I8973bde0113eb1e7876465077605f08e57c45030