Hide FilterComponent match count for CPU profiler
At the moment, the match count isn't accurate in all cases.
(We are using call chart match counts for all four chart
types).

This value should likely be reshown later, but for now we
update the API for FilterComponent to allow using it without
showing a match count at all.

Change-Id: Ifaf1f2eea7eacf1df37314c5416ef2e476d7e51c
Fixes: 112703945
Test: Added