Display fragments in Activity bar tooltip
Users currently have no way to see which fragments
are active. This CL add them in a tooltip that
appears when user hovers over an activity bar in
the event profiler.

There will be more visual changes to come later
to indicate when fragment start and end, but
this pass is at least more useful than nothing.

Bug: 112592038
Test: EventActivityTooltipViewTest updated
Change-Id: I7fee8744c7fed3a23d24ddf1573cfe0f20c7ea6b