Getting ActionType in context
Previously ActionType was a fixed property of an action. However, that
wasn't especially useful in many contexts, since the way an action is
displayed sometimes depends on how the action is viewed (specifically in
the case of an action going to the sibling of a (grand-)parent.
This allows for a tiny bit of cleanup now, but will be more useful once
all actions are represented by SceneComponents.

Test: Updated existing tests
Bug: related to 80325197
Change-Id: I81b4975b28dcb52c9c86c051cb605557faea0800