[Gradle Sync] Reduce the chance of sync destroying the undo stack
This change ensures that Project#save is run in its own undo transparent
action, allowing all file actions to be grouped into the same action
group. As a result there is a lower chance that the global undo stack
will overflow and lose actions. It can still occur however if the
virtual file system events are picked up by one of intellijs jobs
instead of this Project#save.

Bug: 80441694, 112746354
Test: N/A
Change-Id: I4d081df0f10757b105eb77b981613e60acf7d5a1