Fixes for SDK operations cancellation.
Progress indicators did not delegate to each other
appropriately in some contexts, which led to a situation
when a task could be cancelled from foreground wizard,
but not from the IJ background tasks pool. It would say
'Stopping...' until the full completion of the task,
which is not helpful, especially during downloads.

Fixes: b/112721282
Test: N/A
Change-Id: Ic61a10b5489daad536c780a8880d3d43fdd93d15