Refactor CpuThreadsView panel sizing logic
CpuProfilerStageView was using two different layout sizing rules for
CpuThreadsView when it is expanded vs collapsed. However, there is no
visual effect of this logic. Because, Details view has exactly two
children: CpuKernelsView and CpuThreadsView. CpuKernelsView is always
"fit", so there is no difference between "*" or "fit" of the second
children.

Test: existing tests
Bug: b/112682804
Change-Id: If6126c2824b3d13e39a6391b4471ba8df7acfcfc