[CS] Implement basic compound sync support
This CL implements basic support for Compound Sync in the IDE, by
consuming changes in the Model Builder.

If compound sync is enabled (and also single variant sync), IDE will
call model and source generation in a single invocation. This is
implemented for both VariantOnlySync (when changing variants in the UI)
and full sync.

Main changes are:
- Consume new API in SyncExecutor
- Create logic to invoke right option in NewGradleSync
- Introduce new logic in SyncResultHandler and introduce new listener
method

Test: existing should pass, new unit and integration tests added
Bug: 110973454
Change-Id: Iad156fb34d10f0d2f0223adc0996315bbe4c75e0