Remove SampleDataResourceRepository from cache on disposal
SampleDataRepositoryManager caches instances of
SampleDataResourceRepository by the facet that they are associated with.
Currently, we assume that the lifetime of the repository is the same as
the lifetime of its associated facet, but this isn't true since the
repository gets disposed every sync and forced layout refresh. The facet
isn't necessarily disposed in these cases, and so we end up with stale
sample data in the layout editor because we notify the cached repository
of sample data changes and it doesn't update itself because it's already
disposed.

This CL resolves the issue by adding logic which removes a
SampleDataResourceRepository from the cache on disposal, which permits
SampleDataRepositoryManager to construct a fresh instance the next time
it's asked for one.

Change-Id: If4d84eca5afe18706cd8642f90c27ceddbbe3005
Fixes: 112838662
Test: SampleDataResourceRepositoryTest