Shape operation bug fix
A possibility would be for the designer to apply an operation to a
shape, changing its booleanOperation value to 1, 2, 3 or 4, but after a
while decide that the operation is not needed anymore and is therefore
removed. This will change the booleanOperation value to -1, case that
was not considered in previous algorithm.
This CL treats the removal of booleanOperation as a shape union, so that
the shapes (even though they are not intersecting) are united and
displayed together at once. Without this change, any shape with a
booleanOperation of -1 is simply ignored and not displayed, because a
shape operation is expected.

Bug: N/A
Test: N/A
Change-Id: I0988ee969203a67c4196f51c999971e45931fb38