Show Studio analytics consent immediately on launch
With the changes in IntelliJ around consent dialogs, the previous notification
requesting the user to opt-in to providing usage statistics had become
a modal dialog that pops-up at some time after launching the IDE more than a
week after first launch.

This change makes the dialog show up immediately on launch if the user has not
previously made an explicit decision to opt-in or out.

This brings the behavior inline with the behavior of IntelliJ ensures that
the user makes a decision early on, allowing us to capture statistics on
early use.

As this change introduces a modal dialog, there is a new system property
that allows tests to explicitly disable the dialog:
 disable.android.analytics.consent.dialog.for.test

This will leave metrics in the opted-out state.
Fixes: 112910037
Test: manual

Change-Id: Iedca501b1f56f25fd5c77341aafe002d76767d9b

Test-Strategy: ui.test.groups=PROJECT_SUPPORT,PROJECT_WIZARD,THEME,EDITING,TEST_FRAMEWORK,DEFAULT