Fix more tests under light R classes
- Use the resolve scope of the right PSI element when finding R classes,
just like the code editor will do.
- Use the light classes service to get all R classes worth looking for
when refactoring a module. This means the refactoring only works when
light classes are used, but that is a valid dependency since we cannot
ship the namespaces feature without light classes anyway.
- Don't "find" the namespaced R class in AndroidResourceUtil if
namespaces are not used in the project. This is to fix all existing
tests that assert there's only one field corresponding to a given
resource. We will have rework those tests to cover the namespaced case
at some point.

Test: existing tests being fixed to pass under IN_MEMORY_R_CLASSES
Bug: 77801019
Change-Id: I9dae93dd21e8fdf13bf396cac80a50f180cd2f87