Avoid using compatible bitmap for IImageFactory
On certain screen configurations, the device configuration will result
in an ARGB bitmap with alpha premultiplied to be created. Layoutlib is
probably not always creating compatible bitmaps so this causes issues
rendering.
Since this bitmap is not directly rendered to screen but actually just
copied into the image pool, we do not get much benefit from using the
compatible configuration.

Fixes: 113113516
Test: Manually tested
Change-Id: Id02dbbbd6e0a7fa5e809e2b591f61ba6babc162a