Shape operation bug refix
If a shape operation is added and then removed from a shape, the
booleanOperation value will become -1, case that for now is treated as a
union between shapes.
If the shapes do not intersect, we indeed want them to be united so that
both are visible and neither is ignored. However, if the shapes do
intersect, the result in sketch can be either a subtraction or an exclusive or,
depending on how much the shapes overlap. Given that the exclusive or
also behaves like a union for shapes that do not intersect, we can use this
operation instead of just the union, which does not cover the cases
where shapes overlap.

Bug: N/A
Test: N/A
Change-Id: Idab497b12104f0e3d88839a5350565ffaa20ac10