Revert "Stop running commonly timing out test on mac."
This reverts commit d173ce19f853e0e8a7869b5448b8f3ef65f1a1ed.

Reason for revert: This is wrong, and we don't want to do it anyway.
It is wrong because it is in test_tags. So it never even disabled it.
We don't want to do it, because we instead just disabled some of the
long running tests off the PSQ with no_test_mac_psq. See bug.

Bug: 113099009
Test: bazel test //tools/adt/idea/android-lang/...

Change-Id: I3adc4abb454489104c02eb1d150faa0377005b9f