[SVS] Create NdkVariant from NativeVariantAbi.
The previous code creates NdkVariant from NativeArtifact,
but with single-variant sync, the variants are now abi based,
each NativeVariantAbi can contain multiple of NativeArtifact.

This change creates NdkVariant from each NativeVariantAbi if
it's in single-variant mode.

Also, refactor the code that populates class fields, to increase
code re-usage. This also improves the setup for Variant-Only
sync, the previous code re-populate the fields for all of
NativeVariantAbi, and the new code only update the fileds for
the newly added NativeVariantAbi.

Bug: 79581802
Test: Current tests.
Change-Id: Iebe12948b7286799c8a1db7cf6867ead47837ca2