Improve network details editor settings
This CL fixes several misc. UX issues with
IntellijDataViewer classes, as currently the
details editor feels like it's provided an editable
area to the user, instead of a read-only pane.

After this change, for example, editor margins
are removed, as are refactoring / analyzing context
menu options.

Also, I added some long missing tests to verify
behavior of this (and other) IntellijDataViewer
functionality.

Fixes: 112471583
Fixes: 112472387
Fixes: 112471434
Fixes: 113209974
Test: Almost full coverage added to IntellijDataViewer (minus one
exception block we can't easily trigger in test)

Change-Id: Ib7cf12b26baba1c0db4bfdbd9b2433f94c2151a7