Transparency bug fix
Currently, the engine only checks for gradient overall transparency
and applies it to the shape.
On top of that, however, the whole layer
containing the shapeGroup can have it's own transparency settings.
Moreover, the shape in question can be included in a group that has its
own transparency settings. All this opacity must be accumulated from
the root to the last node and applied appropriately. This is what thie
CL does.

Just like the parent coordinates keep adding up as we go deeper into the
tree, also does the opacity until we reach a shapegroup. The final
opacity is then passed in the ShapeModel constructor, where the opacity
is applied on the SketchStyle field of the object.

Bug: N/A
Test: DrawableFileGeneratorTest.java
Change-Id: Iccd04903a7e6c128e94d126a85af908b85a8ef22