Have SampleDataListener use file extension instead of type
SampleDataListener#isPossiblyRelevant uses VirtualFile#getFileType
to "quickly"(TM) detect and ignore Java and XML file changes, which
aren't relevant to sample data. It turns out that getFileType() parses
the file to determine this information, which makes this check very
slow and causes UI freezing.

This CL resolves the issue by having isPossiblyRelevant() check the
extension of the file rather than figuring out the file type. This
method is a little less precise, but should actually be fast.

Change-Id: If6b76b43eb5d42acefede8655e4045fbff007c6a
Fixes: 113335081
Test: manual