Remove obsolete layout code from HttpDataComponentFactory
It seems there was some old code which was involved in laying
out components in the network request / response tabs, forcing
a minimum height because sometimes images wouldn't show (or at
least, that's the best reason I could guess from the comments).

At this point, a few other things have changed - we use a new
LayoutManager in this UI, and we are also better about making
sure that every component type we render in this area has a
valid preferred size set.

Thanks to these changes, when I tried to run latest with the
"minimum layout" logic removed, displaying many different image
and text formats worked fine.

It's also nice to remove this code because it's kind of a hack,
branching special-case logic for images.

Bug: N/A
Test: N/A, removed code only
Change-Id: I9ac171045e27f1f50d5aa11cf588999e6fd34ed9