[New Sync] Fix ndk module setup from cache.
NDK modules have both of AndroidModuleModel and NdkModuleModel,
When setting up modules from cache, check existence of NdkModuleModel
when AndroidModuleModel is not empty. The original code returns
immediately when AndroidModuleModel is found, resulting in
NdkModuleModel not being setup.

Update the test to make NDK module contain both of AndroidModuleModel
and NdkModuleModel.

Bug: NA
Test: CachedProjectModelsSetupTest
Change-Id: Ic028f912c1163817840a2879fd6b7c6f281104cd